Created Gameboard and Cell function
The gameboard contains row and column and a for
loop to build 2D array on every iteration used
push method and call Cell function to return
the object with two method addToken and getValue,
and a private variable value.

The Cell function has two method to add a token
based on the player value which is either 0, X,
or O. Also get value return the value of the cell
at each point.
[Ticket: 1]