Create dropToken function
in the Gameboard function I added a new method
called dropToken which take three values as
row, column and player. then using board array
it will targer specific place and put the player
token in the spot. if the spot is already full,
then it prints the message that the spot is not
empty.

[Ticket: 2]