Build gameController function
this is where the player array is created. inside
each one there is an object with name and token
attributes. then activePlayer is defined based on
the current player. Also switchPlayer will toggle
players on each round. At the end playRound
function is defined to place the mark of each
player (token) either "X" or "O" on the board and
finally print the board on the console.

[Ticket: 4]