Add reset functionality to the game.
When reset ic clicked, three actions need to
happen. 1. the board is set back to the original
status. with addeventlistener and for loop to
create new 2D array and assign it to board we
can achive this. first Gameboard needs to be
called and assigned to variable gameboard, and
then method gatBoard is called and set it to
board variable. now we have access to the board
that we have been playing with to set it back
to the original form.
2. all cells has to be cleared
3. form needs to be cleared and active again

[Ticket: 7]