Fix the bug in board
The issue was when the game is reset, the board
did not become empty. there was lack of connection
the reason was the board is recreated at
displayGame which was wholy different board.
To fix it I moved the reset functionality inside
gameController and call the method cleanBoard of
the Gameboard, so that the same board is modified.

[Ticket: 8]